математика

+++ src/main/java/got/client/gui/GOTGuiButtonOptions.java
@@ -27,7 +27,7 @@ public class GOTGuiButtonOptions extends GuiButton {
 			GL11.glEnable(3042);
 			GL11.glBlendFunc(770, 771);
 			Gui.func_146110_a(xPosition, yPosition, 170.0f, 316.0f, width / 2, height, 512.0f, 512.0f);
-			Gui.func_146110_a(xPosition + width / 2, yPosition, 370 - width / 2, 316.0f, width / 2, height, 512.0f, 512.0f);
+			Gui.func_146110_a(xPosition + width / 2, yPosition, 370 - (float) width / 2, 316.0f, width / 2, height, 512.0f, 512.0f);
 			mouseDragged(mc, i, j);
 			int color = 8019267;
 			if (!enabled) {
